fix: InputView 상태 분리로 인한 class 추가
