+++ src/main/java/lotto/view/ManualCountInputView.java
@@ -9,23 +9,11 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
 
-public class InputView {
+public class ManualCountInputView {
     public static final int UNIT_PRICE = 1000;
     public static final String DELIMITER = ",";
 
-    private int money;
     private int manualLottoCount;
-    private List<Integer> winNumbers;
-    private int bonusNumber;
-
-    public void setMoney() {
-        System.out.println("구입금액을 입력해 주세요.");
-        this.money = new Scanner(System.in).nextInt();
-    }
-
-    public int getMoney() {
-        return this.money;
-    }
 
     public void setManualLottoCount() {
         System.out.println("수동으로 구매할 로또 수를 입력해 주세요.");
@@ -57,27 +45,5 @@ public class InputView {
         for (Lotto lotto : lottos) {
             System.out.println(lotto);
         }
-    }
-
-    public void setWinNumbers() {
-        System.out.println("지난 주 당첨 번호를 입력해 주세요.");
-        this.winNumbers = Arrays.stream(new Scanner(System.in).nextLine()
-                .split(DELIMITER))
-                .map(String::trim)
-                .map(Integer::parseInt)
-                .collect(Collectors.toList());
-    }
-
-    public void setBonusNumber() {
-        System.out.println("보너스 볼을 입력해 주세요.");
-        this.bonusNumber = new Scanner(System.in).nextInt();
-    }
-
-    public List<Integer> getWinNumbers() {
-        return winNumbers;
-    }
-
-    public int getBonusNumber() {
-        return bonusNumber;
     }
 }
